refactor(ProductBaseViewHolder): ViewHolder 추상화
