+++ app/src/main/java/woowacourse/shopping/presentation/productlist/product/RecentProductContainerViewHolder.kt
@@ -1,9 +1,10 @@
-package woowacourse.shopping.presentation.productlist.recentproduct
+package woowacourse.shopping.presentation.productlist.product
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemRecentProductContainerBinding
 import woowacourse.shopping.presentation.model.ProductModel
 import woowacourse.shopping.presentation.model.ProductViewType.RecentProductModels
+import woowacourse.shopping.presentation.productlist.recentproduct.RecentProductAdapter
 
 class RecentProductContainerViewHolder(
     private val binding: ItemRecentProductContainerBinding,
