+++ app/src/main/java/woowacourse/shopping/presentation/productlist/recentproduct/RecentProductItemViewHolder.kt
@@ -1,30 +1,17 @@
 package woowacourse.shopping.presentation.productlist.recentproduct
 
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ItemRecentProductBinding
 import woowacourse.shopping.presentation.model.ProductModel
+import woowacourse.shopping.presentation.productlist.product.ProductBaseViewHolder
 
 class RecentProductItemViewHolder(
     private val binding: ItemRecentProductBinding,
-    private val showProductDetail: (ProductModel) -> Unit,
-) : RecyclerView.ViewHolder(binding.root) {
+    showProductDetail: (ProductModel) -> Unit,
+) : ProductBaseViewHolder(binding.root, showProductDetail) {
 
-    private var _productModel: ProductModel? = null
-    private val productModel get() = _productModel!!
-
-    init {
-        itemView.setOnClickListener { showProductDetail(productModel) }
-    }
-
-    fun bind(recentProduct: ProductModel) {
+    override fun bind(recentProduct: ProductModel) {
         _productModel = recentProduct
         binding.textProductListName.text = recentProduct.name
-        Glide.with(binding.imageProductListPoster.context)
-            .load(recentProduct.imageUrl)
-            .error(R.drawable.default_image)
-            .centerCrop()
-            .into(binding.imageProductListPoster)
+        setImage(recentProduct, binding.imageProductListPoster)
     }
 }
